feat: Add new icons for Github and Play
This commit adds new icons for the Github and Play components. The Github icon is a new SVG file added to the `src/icons` directory, while the Play icon is also a new SVG file added to the same directory. These icons will be used to enhance the user interface and provide visual cues for important actions.