refactor: Update TranslateImage.astro with improved file input accept attribute
This commit updates the TranslateImage.astro file to improve the file input element. The accept attribute of the input element has been modified to accept only PNG and JPG image files. This change ensures that only valid image files can be selected for translation, enhancing the accuracy of the translations and providing a better user experience.